Fix high CPU usage, add winning/losing screen
