+++ main.py
@@ -15,8 +15,8 @@ def main():
             multichoice = -1
             while multichoice != 3:
                 multichoice = wrapper(multiplayer_menu)
-                if multichoice == 1: host(wrapper(multiplayer_address_menu, 0, "127.0.0.1", "4444"), grid)
-                if multichoice == 2: join(wrapper(multiplayer_address_menu, 1, "88.160.115.96", "25565"))
+                if multichoice == 1: host(wrapper(multiplayer_address_menu, 0, "127.0.0.1", "25565"), grid)
+                if multichoice == 2: join(wrapper(multiplayer_address_menu, 1, "127.0.0.1", "25565"))
         
 if __name__ == "__main__":
     main()
